+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # data_tools version history
 
 ## v0.0.8 (WIP):
+- In `plots` module:
+    - Function `similarity_heatmap`: Given a group of sets, generates a
+      heatmap with the similarity indices across each possible pair.
+    - Function `similarity_histogram`: Given a group of sets, generates
+      a histogram of the similarity indices across each possible pair (same-element pairs excluded).
 
 ## v0.0.7:
 - Support for Python 3.x versions.